completely change the strategy for inlining jit_merge_points in the caller. The old one did not work in complex cases involving raising graphs. In particular, this case did not work:
def bar():
    "something which cannot be inlined and raises"

@inline_in_portal
def foo():
    driver.jit_merge_point()
    return bar()

def fn():
    try:
        foo():
    except StopIteration:
        pass

that's because that the backendopt inliner is not able to inline calls to
raising graphs inside a try/except block.

To work around the issue, we put the actual driver.jit_merge_point in a
separate function, which needs to be called as soon as we enter foo(). Then,
we move *only* this call from foo() to fn(), and finally inline the
jit_merge_point in fn().

Next step is to provide a nice decorator to do everythin automatically

--HG--
branch : autoreds