fix test_guard_not_forced_fails by refactoring the way libffi_call works. The problem is that we need to make sure that the box containing the result of call_release_gil is considered alive by the codewriter when we write it to the exchange_buffer.
The solution is to create a new separate operation libffi_save_result whose
job is to write the result in the exchange_buffer: during normal execution
this is a no-op because the buffer is already filled by libffi, but during
jitting the behavior is to actually write into the buffer.

The result is that now the jitcode looks like this:

%i0 = libffi_call_int(...)
-live- libffi_save_result_int(..., %i0)

the "-live-" is the key, because it make sure that the value is not lost if
guard_not_forced fails.

Note that rlib/jit_libffi.py has not been updated yet, will do in the next
commits

--HG--
branch : refactor-call_release_gil